# docs/openapi.json
{
  "components": {
    "schemas": {
      "BalanceOut": {
        "properties": {
          "area_id": {
            "title": "Area Id",
            "type": "integer"
          },
          "area_name": {
            "title": "Area Name",
            "type": "string"
          },
          "item_base_unit": {
            "title": "Item Base Unit",
            "type": "string"
          },
          "item_id": {
            "title": "Item Id",
            "type": "integer"
          },
          "item_name": {
            "title": "Item Name",
            "type": "string"
          },
          "qty": {
            "title": "Qty",
            "type": [
              "number",
              "null"
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "item_name",
          "item_base_unit",
          "area_id",
          "area_name",
          "updated_at"
        ],
        "title": "BalanceOut",
        "type": "object"
      },
...
      "CountAdjustmentIn": {
        "properties": {
          "area_id": {
            "title": "Area Id",
            "type": "integer"
          },
          "item_id": {
            "title": "Item Id",
            "type": "integer"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              }
            ],
            "description": "Signed compensating qty in base units. Human-driven only \u2014 never auto-derived from unexplained variance (D13).",
            "title": "Qty"
          }
        },
        "required": [
          "item_id",
          "area_id",
          "qty"
        ],
        "title": "CountAdjustmentIn",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "MovementOut": {
        "properties": {
          "area_id": {
            "title": "Area Id",
            "type": "integer"
          },
          "area_name": {
            "title": "Area Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "item_id": {
            "title": "Item Id",
            "type": "integer"
          },
          "item_name": {
            "title": "Item Name",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "qty": {
            "title": "Qty",
            "type": [
              "number",
              "null"
            ]
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "item_id",
          "item_name",
          "area_id",
          "area_name",
          "kind",
          "source_type",
          "source_id",
          "occurred_at",
          "note"
        ],
        "title": "MovementOut",
        "type": "object"
      },
...
      "TransferIn": {
        "properties": {
          "from_area_id": {
            "title": "From Area Id",
            "type": "integer"
          },
          "item_id": {
            "title": "Item Id",
            "type": "integer"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              }
            ],
            "description": "Positive amount to move (base units).",
            "title": "Qty"
          },
          "to_area_id": {
            "title": "To Area Id",
            "type": "integer"
          }
        },
        "required": [
          "item_id",
          "from_area_id",
          "to_area_id",
          "qty"
        ],
        "title": "TransferIn",
        "type": "object"
      },
      "TransferOut": {
        "properties": {
          "transfer_in": {
            "$ref": "#/components/schemas/MovementOut"
          },
          "transfer_out": {
            "$ref": "#/components/schemas/MovementOut"
          }
        },
        "required": [
          "transfer_out",
          "transfer_in"
        ],
        "title": "TransferOut",
        "type": "object"
      },
...
      "WasteIn": {
        "properties": {
          "area_id": {
            "title": "Area Id",
            "type": "integer"
          },
          "item_id": {
            "title": "Item Id",
            "type": "integer"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              }
            ],
            "description": "Positive amount wasted (base units). Stored as negative movement.",
            "title": "Qty"
          }
        },
        "required": [
          "item_id",
          "area_id",
          "qty"
        ],
        "title": "WasteIn",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend contract for Evolving Cook. Auth: Bearer token from /auth/login (Django signing, 30 days). 401 is never a reason to drop client data \u2014 re-auth and retry. Walk batch submit 401: re-auth and retry the batch; never discard line payloads. Phase 1.5: /boards/* MEP service boards (covers never required). Phase 2: /walks/* lifecycle + /purchasing/* PO draft\u2192send + delivery receipt. Phase 3: /inventory/* append-only stock ledger (balances, movements, waste, count_adjustment, transfer). Delivery complete \u2192 receipt movements; walk submit/lock \u2192 theoretical + unexplained variance (never auto-correct). Phase 4: /boards/* full planner \u2014 banquet BEO covers + waves + WaveAllocation (one line across waves), produce scale formula proposed_qty = scaling_covers * template.yield_per_cover (null yield \u2192 null proposed), canteen produce qty entered (not covers-derived), day close + outturn. Covers never required outside banquet*. D12/D14: line notes (today) + template_notes (dish template, every day) + item notes/house_made. NOTE\u2192ASSIST v2: note-save auto-enqueues parse_note; board lines carry pending_proposal inline; accept target line|template|item. Phase 5: /assist/* AssistProposal accept/reject + jobs (Hermes A2A); B15 explode via POST /assist/explode. Internal agent-events is NOT on /api/v1.",
    "title": "Evolving Cook API",
    "version": "0.1.13"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inventory/balances": {
      "get": {
        "operationId": "api_inventory_get_balances",
        "parameters": [
          {
            "in": "query",
            "name": "item_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Item Id"
            }
          },
          {
            "in": "query",
            "name": "area_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BalanceOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List stock balances (filter item and/or area)",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/inventory/count-adjustment": {
      "post": {
        "operationId": "api_inventory_post_count_adjustment",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountAdjustmentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovementOut"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Human compensating count_adjustment movement. Variance is unexplained only \u2014 never auto-posts adjustments.",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/inventory/movements": {
      "get": {
        "operationId": "api_inventory_get_movements",
        "parameters": [
          {
            "in": "query",
            "name": "item_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Item Id"
            }
          },
          {
            "in": "query",
            "name": "area_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Id"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MovementOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List recent stock movements (filter item/area/kind)",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/inventory/transfer": {
      "post": {
        "operationId": "api_inventory_post_transfer",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferOut"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Paired transfer_out + transfer_in in one transaction",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/inventory/waste": {
      "post": {
        "operationId": "api_inventory_post_waste",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WasteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovementOut"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Post waste movement (balance decreases)",
        "tags": [
          "inventory"
        ]
      }
    },
...
  },
  "servers": []
}